+++ ForlornWreckageStage5/Thunderstore/CHANGELOG.md
@@ -1,4 +1,14 @@
 <details>
+<summary>1.0.2 </summary>
+
+* Removed Access Node spawn. Added config to re-enabled the spawn.
+</details>
+<details>
+<summary>1.0.1 </summary>
+
+* Fixed stage order (so it now appears at appropriate time in bazaar seers).
+</details>
+<details>
 <summary>1.0.0 </summary>
 
 * Initial release